analyst-rating-simple/src/requirements.txt
basebot22-basebot-JustinGuese
pandas
numpy
requests